==> 25-tomokiyo/01_control_group_matching_process/01_extract_embedding_eid.ipynb <==
## Step1: Specialization Vector

## Read Scopus all embedding data

import pickle
import pandas as pd
import numpy as np

# Read embedding data of papers generated by LLM
embedding_df = pd.read_pickle('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/scopus/who_scopus_candidate_embedding_2011_2015.pickle')


embedding_df.head()

embedding_df.shape

embedding_eid_ls = embedding_df['eid'].tolist()

# np.save('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/embedding_eid_ls.npy',embedding_eid_ls)
