==> 25-tomokiyo/01_control_group_matching_process/02_additional_embedding.ipynb <==
## Step1: Specialization Vector

## For data with eid in Scopus but with blank title or abstract, embedding was performed by checking the original paper

import pandas as pd

title_df = pd.read_pickle('/disks/qnap3/shared/scopus-24/data/paper/title.pickle')
title_df = pd.DataFrame(title_df)
title_df.reset_index(inplace=True)
title_df.rename(columns={'index':'eid'},inplace=True)

abstract_df = pd.read_pickle('/disks/qnap3/shared/scopus-24/data/paper/abstract.pickle')
abstract_df = pd.DataFrame(abstract_df)
abstract_df.reset_index(inplace=True)
abstract_df.rename(columns={'index':'eid'},inplace=True)

abstract_df[abstract_df['eid']==84875442265]

- Patdu_Kaye
    - "Co-benefits: Taking a multidisciplinary approach", 84875442265, NO ABSTRACT, 2013

- Whaley_Paul
  - "Priorities for cancer prevention", 84861568429, NO ABSTRACT, 2012
  - "Childhood obesity and the environment", 85026817599, NO ABSTRACT, 2014



import pandas as pd, os
import numpy as np
import time
from multiprocessing import Pool
flatten2 = lambda l: [item for sublist in l if sublist == sublist for item in sublist]
import os,time

import time
import pandas as pd
import numpy as np
from openai import OpenAI

# os.environ["OPENAI_API_KEY"] = "##"

client = OpenAI()

def get_embeddings(texts, model="text-embedding-3-small", dimensions=None):
    try:
        time.sleep(1)
        texts = [text.replace("\n", " ") for text in texts]
        
        if dimensions:
            res = client.embeddings.create(input=texts, model=model, dimensions=dimensions)
        else:
            res = client.embeddings.create(input=texts, model=model)
        
        emb = [x.embedding for x in res.data]
        df = pd.DataFrame(np.stack(emb))
        return df
    
    except Exception as e:
        print(f"Error: {e}")
        return pd.DataFrame()

title_abstract_ls = [
    "Co-benefits: Taking a multidisciplinary approach. Most environmental policies are meant to achieve a single goal such as reducing pollution, preserving ecosystems or mitigating climate change. Yet environmental issues are frequently interlinked and approaches that exploit these linkages can deliver more than one desirable outcome. The collection of benefits accruing to actions linking climate change and other development priorities (e.g., reducing local pollution) are known as ‘co-benefits’ Citation[1]. Over the past two decades, a growing number of studies have demonstrated that an integrated or co-benefits approach could prove more cost effective than managing climate and development issues in isolation, but policies reflecting this integration have tended to lag behind research. This article argues that a multidisciplinary approach will be needed to bring policies in line with research on co-benefits.", 
    "Priorities for cancer prevention. Bernard Stewart's Personal View1 in favour of lifestyle choices to prevent cancer seems to be based on the questionable premise that pollutants should be investigated individually for their ability to cause disease. This approach might work for tobacco smoke and other easily isolated causes with obvious effects that lend themselves well to epidemiological study. However, it is not suitable for variable inadvertent exposures to complex mixtures of substances for long periods. Our best understanding is that several carcinogenic exposures accumulate and the overall effect is what causes cancer in individuals and populations.",
    "Childhood obesity and the environment. The rising rate of infant obesity presents the biggest challenge to diet- and lifestyle-based explanations for the obesity epidemic. Although the hypothesis that everyday exposure to chemicals may contribute to risk of obesity is controversial, there is pharmaceutical proof of concept from diabetes drugs in humans and diethylstilboestrol in a mouse model. Chemical exposure also complicates the role of obesity in the etiology of diabetes and cancer. Given that no chemical has been tested for obesogenicity before being brought to market, it is overly hasty to dismiss the obesogen hypothesis as a “green scare."
]

embedding_df = get_embeddings(title_abstract_ls)

embedding_df['eid'] = ["84875442265","84861568429","85026817599"]

# embedding_df.to_pickle('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/additional_embedding_2011-2015.pickle')

embedding_df
