==> 25-tomokiyo/01_control_group_matching_process/03_create_one_hot_vector.ipynb <==
## Step1: Specialization Vector

## Create a one-hot vector for each paper indicating whether or not each author wrote it
## This one-hot vector is used for creating specialization vectors of authors

import pickle
import pandas as pd
import numpy as np

# Read authids in each section of the guideline
authid_12 = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/member_data/2021/A1.2_who_guideline_development_group.npy')
authid_13 = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/member_data/A1.3_who_systematic_review_team_revise.npy')
authid_14 = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/member_data/2021/A1.4_who_external_methodologists.npy')
authid_15 = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/member_data/A1.5_who_external_review_team_revise.npy')

boradmember_authid = list(authid_12)+list(authid_13)+list(authid_14)+list(authid_15)

additional_embedding_df = pd.read_pickle('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/additional_embedding_2011-2015.pickle')
additonal_embedding_eid = additional_embedding_df['eid'].values
additonal_embedding_eid = additonal_embedding_eid.astype(int)
additonal_embedding_eid

authid_list = [35780677200 if x == 7006806977 else 57206422848 if x == 54791745100  else 57219144066 if x == 55588756600  else 57219144066 if x == 57191727610  else 57219144066 if x == 14037485900 else x for x in boradmember_authid]

embedding_eid_ls = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/embedding_eid_ls.npy')
embedding_eid_ls1 = list(embedding_eid_ls) + list(additonal_embedding_eid)

author_paperlist_df = pd.read_pickle('/disks/qnap3/shared/scopus-24/data/author/author_paperlist.pickle')
author_paperlist_df = pd.DataFrame(author_paperlist_df)

author_paperlist_df.head()

author_paperlist_df_explode = author_paperlist_df.explode('author_paperlist')

author_paperlist_df_explode.rename(columns={'author_paperlist':'eid'},inplace=True)

author_paperlist_df_explode.reset_index(inplace=True)

author_paperlist_df_explode1_1 = author_paperlist_df_explode[author_paperlist_df_explode['eid'].isin(embedding_eid_ls1)]

authid_list = [35780677200 if x == 7006806977 else 57206422848 if x == 54791745100  else 57219144066 if x == 55588756600  else 57219144066 if x == 57191727610  else 57219144066 if x == 14037485900 else x for x in boradmember_authid]
authid_list = list(set(authid_list))
len(authid_list)

author_paperlist_df_explode1_1.replace({7006806977:35780677200, 54791745100:57206422848, 55588756600:57219144066, 57191727610:57219144066, 14037485900:57219144066},inplace=True)

# author_paperlist_df_explode1.head(50).to_csv('create_one_hot_for_gpt.csv')

author_ls = author_paperlist_df_explode1_1.authids.unique()

import pandas as pd
import numpy as np
from scipy.sparse import csr_matrix

df = author_paperlist_df_explode1_1
unique_authids = df["authids"].unique()
unique_eids = embedding_eid_ls1

authid_to_idx = {authid: idx for idx, authid in enumerate(unique_authids)}
eid_to_idx = {eid: idx for idx, eid in enumerate(unique_eids)}

row_indices = df["authids"].map(authid_to_idx).values
col_indices = df["eid"].map(eid_to_idx).values

# sparse_matrix: one-hot encoding that represents the presence of each authid in each eid
sparse_matrix = csr_matrix((np.ones(len(df), dtype=np.int8), (row_indices, col_indices)), 
                           shape=(len(unique_authids), len(unique_eids)))

# Save file
# output_file = "/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/sparse_one_hot_matrix.npz"
#np.savez_compressed(output_file, data=sparse_matrix.data, indices=sparse_matrix.indices, indptr=sparse_matrix.indptr, shape=sparse_matrix.shape)
#print(f"Sparse Matrix Shape: {sparse_matrix.shape}")
#print(f"Non-zero Elements: {sparse_matrix.nnz}")
#print(f"Saved File Path: {output_file}")
