==> 25-tomokiyo/01_control_group_matching_process/04_dataframe2ndarray.ipynb <==
## Step1: Specialization Vector

## Tramsform embedding dataframe to ndarray

import pickle
import pandas as pd
import numpy as np

embedding_df = pd.read_pickle('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/scopus/who_scopus_candidate_embedding_2011_2015.pickle')

embedding_df.head(2)

additional_embedding_df = pd.read_pickle('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/additional_embedding_2011-2015.pickle')

additional_embedding_df.head(2)

embedding_df = pd.concat([embedding_df, additional_embedding_df])

embedding_df.reset_index(drop=True, inplace=True)

embedding_df.head(1)

embedding_df.drop(columns=['eid'],inplace=True)

dense_matrix = embedding_df.values 

dense_matrix.shape

# np.save('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/embedding_matrix_2011_15.npy', dense_matrix)
