==> 25-tomokiyo/01_control_group_matching_process/05_cal_dot.ipynb <==
## Step1: Specialization Vector

## Take the dot product of the paper vector and the researcher's one-hot vector for creating the specialization vector

import pickle
import pandas as pd
import numpy as np
from scipy.sparse import csr_matrix

# dense_matrix: the matrix of the paper embedding vectors
dense_matrix = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/embedding_matrix_2011_15.npy')

file_path = "/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/sparse_one_hot_matrix.npz"
loaded_data = np.load(file_path)
# sparse_matrix: one-hot vector for each paper indicating whether or not each author wrote it
sparse_matrix = csr_matrix((loaded_data["data"], loaded_data["indices"], loaded_data["indptr"]),
                           shape=loaded_data["shape"])

# Dot product
result_matrix = sparse_matrix.dot(dense_matrix)  # (11289650, 1536)
researcher_paper_counts = sparse_matrix.sum(axis=1)  # (11289650, 1)
researcher_paper_counts = np.array(researcher_paper_counts).flatten()
researcher_paper_counts[researcher_paper_counts == 0] = 1 
average_researcher_vectors = result_matrix / researcher_paper_counts[:, None]  # (11289650, 1536)
print(f"Average Researcher Vectors Shape: {average_researcher_vectors.shape}")  # (11289650, 1536)

# average_researcher_vectors: specialization vector for each researcher
average_researcher_vectors.shape

# np.save('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/researcher_average_embedding_matrix_2011_15.npy', average_researcher_vectors)
