==> 25-tomokiyo/01_control_group_matching_process/07_matching_mse.ipynb <==
## Step2: Specialization Similarity Ranking
## Step3: Matched Control Researcher

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
pd.options.display.max_rows = 100
pd.options.display.max_colwidth = None
import pickle
from tqdm import tqdm
import seaborn as sns
import os, sys, gc, random, itertools, math, pickle
from itertools import combinations
import networkx as nx


# Read Author IDs of the guideline members
authid_12 = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/member_data/2021/A1.2_who_guideline_development_group.npy')
authid_13 = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/member_data/A1.3_who_systematic_review_team_revise.npy')
authid_14 = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/member_data/2021/A1.4_who_external_methodologists.npy')
authid_15 = np.load('//disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/member_data/A1.5_who_external_review_team_revise.npy')

boradmember_authid = list(authid_12)+list(authid_13)+list(authid_14)+list(authid_15)

authid_list1 = [35780677200 if x == 7006806977 else 57206422848 if x == 54791745100  else 57219144066 if x == 55588756600  else 57219144066 if x == 57191727610  else 57219144066 if x == 14037485900 else x for x in boradmember_authid]
treatment_authids = list(set(authid_list1))

candidate_feature_value = pd.read_pickle('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/candidate_feature_value.pickle')

candidate_feature_value

top_k_similar_authids = np.load('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/top_k_similar_authids.npy', allow_pickle=True)
top_k_similar_authids = top_k_similar_authids.item()

# Read top_k_similar_authids and remove treatment_authids
new_top_k_similar_authids_dict = {}

for index, top_similar_authid in zip(top_k_similar_authids.keys(), top_k_similar_authids.values()):
    tmp = []
    for authid in top_similar_authid:
        if authid not in treatment_authids:
            tmp.append(authid)
    new_top_k_similar_authids_dict[index] = tmp
            

## Focus on the top 20 similar authors for each author
top_k_similar_authids_dict = {}

for index, top_similar_authid in zip(new_top_k_similar_authids_dict.keys(), new_top_k_similar_authids_dict.values()):
    tmp = []
    for authid in top_similar_authid:
        tmp.append(authid)
    top_k_similar_authids_dict[index] = tmp[:20]

candidate_feature_value.head()

import pandas as pd
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
from scipy.spatial.distance import cdist

# Extracting guideline researchers
guideline_researchers = treatment_authids

# List to store the matched control candidates
matched_control_candidate_group = {}

matched_control_group = []

used_treatment_authids = []

for guideline_researcher in tqdm(treatment_authids):

    if guideline_researcher in top_k_similar_authids_dict.keys():

        used_treatment_authids.append(guideline_researcher)
    
        top_n_candidate_authid = top_k_similar_authids_dict[guideline_researcher]
        candidate_researcher = list(top_n_candidate_authid)+list([guideline_researcher])
    
        # Create a DataFrame with the candidates
        yearly_num_cited_df = candidate_feature_value[candidate_feature_value.index.isin(candidate_researcher)]
        treat_citation_history = yearly_num_cited_df[yearly_num_cited_df.index==guideline_researcher][[f'academic_citation_{i}' for i in range(2011,2016)]+['policy_citation_2015','paper_num_2015']]
        treat_citation_history = treat_citation_history.astype(float)
        
        control_candidate_citation_history = yearly_num_cited_df.loc[top_n_candidate_authid][[f'academic_citation_{i}' for i in range(2011,2016)]+['policy_citation_2015','paper_num_2015']]
        control_candidate_citation_history = control_candidate_citation_history.astype(float)
        
        treat_citation_history_series = treat_citation_history.loc[guideline_researcher].tolist()
        diff = control_candidate_citation_history.sub(treat_citation_history_series, axis='columns') #academic_citation_2011-2015とpolicy_citation_2015で平均二乗誤差
        mse = np.sqrt((diff**2).sum(axis=1))
        
        # Sort the candidates based on the MSE
        n = 10
        similar = mse.sort_values(ascending=True)[0:n].index.values
    
        for index, candidate in enumerate(list(similar)):
            if candidate not in matched_control_group:
                matched_control_group.append(candidate)
                break
                
            
      
        #plt.plot(yearly_num_cited_df.loc[guideline_researcher,['academic_citation_2011','academic_citation_2012','academic_citation_2013','academic_citation_2014','academic_citation_2015','policy_citation_2015','paper_num_2015']], color="r", linewidth=5,alpha=0.5, label="Nobel Laureates")
        
        #for i in list(similar):
            #plt.plot(yearly_num_cited_df.loc[i,['academic_citation_2011','academic_citation_2012','academic_citation_2013','academic_citation_2014','academic_citation_2015','policy_citation_2015','paper_num_2015']], color="b",alpha=0.5)
        #plt.plot(yearly_num_cited_df.loc[list(similar)[index],['academic_citation_2011','academic_citation_2012','academic_citation_2013','academic_citation_2014','academic_citation_2015','policy_citation_2015','paper_num_2015']], color="b",alpha=0.5,linewidth=5, label=f"top of {n} similar Researchers")
        
        #plt.xticks(rotation=45)
        #plt.legend()
        #plt.show()
    else:
        continue


yearly_num_cited_df.head()

# Create a DataFrame with the matched treatment and control groups
matched_data = pd.DataFrame()

treatment_data = candidate_feature_value[candidate_feature_value.index.isin(used_treatment_authids)]
control_data = candidate_feature_value[candidate_feature_value.index.isin(matched_control_group)]

matched_data = pd.concat([treatment_data, control_data])

matched_data['guideline_member'] = 0
matched_data.loc[matched_data.index.isin(used_treatment_authids), 'guideline_member'] = 1

matched_data.head()

# matched_data.to_pickle('/disks/qnap3/users/21-tomokiyo/masters_thesis/data/guideline_paper/all_guideline_matching/matched_data.pickle')
